Move `Discipline` to the `character` package
