+++ src/main/kotlin/character/ActiveClassJob.kt
@@ -1,7 +1,5 @@
 package cloud.drakon.ktlodestone.character
 
-import cloud.drakon.ktlodestone.character.profile.Discipline
-
 /** A Character's active [ClassJob]. */
 data class ActiveClassJob(
     val classJob: ClassJob,
