+++ src/main/kotlin/character/profile/ScrapeCharacterProfile.kt
@@ -2,6 +2,7 @@ package cloud.drakon.ktlodestone.character.profile
 
 import cloud.drakon.ktlodestone.character.ActiveClassJob
 import cloud.drakon.ktlodestone.character.ClassJob
+import cloud.drakon.ktlodestone.character.Discipline
 import cloud.drakon.ktlodestone.character.Guild
 import cloud.drakon.ktlodestone.character.grandcompany.GrandCompany
 import cloud.drakon.ktlodestone.character.grandcompany.GrandCompanyName
